refactor BookList in a single component (beta)
