+++ src/app/screens/BookList/index.tsx
@@ -1,22 +1,36 @@
 import React, { useEffect } from 'react';
+import { ListRenderItem, FlatList } from 'react-native';
 import { useSelector, useDispatch } from 'react-redux';
 
 import withLoading from '@components/WithLoading';
 import { AppState } from '@interfaces/appState';
+import { Book } from '@interfaces/book';
 import bookActions from '@redux/books/actions';
-import { BOOK_REDUCER } from '@redux/constants';
 
-import BookList from './layout';
+import BookCard from './components/BookCard';
+import styles from './styles';
 
 function BookListContainer() {
-  const BookListWithLoading = withLoading(BookList);
+  const LoadableFlatlist = withLoading(FlatList);
+  const keyExtractor = ({ id }: Book) => String(id);
   const dispatch = useDispatch();
-  const { books, booksLoading } = useSelector((state: AppState) => state[BOOK_REDUCER]);
-
+  const { books, booksLoading } = useSelector((state: AppState) => state.book);
+  const renderBook: ListRenderItem<Book> = ({ item, index }) => (
+    <BookCard customStyle={index !== books!?.length - 1 && styles.bookListItem} {...item} />
+  );
   useEffect(() => {
     dispatch(bookActions.getBooks());
   }, [dispatch]);
 
-  return <BookListWithLoading loading={booksLoading} books={books} />;
+  return (
+    <LoadableFlatlist<Book>
+      loading={booksLoading}
+      data={books}
+      renderItem={renderBook}
+      keyExtractor={keyExtractor}
+      contentContainerStyle={styles.bookListInside}
+      style={styles.bookList}
+    />
+  );
 }
 export default BookListContainer;
