--- src/redux/constants.ts
@@ -1,2 +1,0 @@
-export const AUTH_REDUCER = 'auth';
-export const BOOK_REDUCER = 'book';
