+++ src/redux/store.ts
@@ -3,13 +3,12 @@ import thunk from 'redux-thunk';
 
 import Reactotron from '@config/reactotron';
 
-import authReducer from './auth/reducer';
-import bookReducer from './books/reducer';
-import { AUTH_REDUCER, BOOK_REDUCER } from './constants';
+import auth from './auth/reducer';
+import book from './books/reducer';
 
 const rootReducer = combineReducers({
-  [AUTH_REDUCER]: authReducer,
-  [BOOK_REDUCER]: bookReducer
+  auth,
+  book
 });
 
 const store = createStore(rootReducer, compose(applyMiddleware(thunk), Reactotron.createEnhancer()));
